requirements.txt
fastapi
openai-whisper
python-multipart
uvicorn